chore(builder): rename payload_builder.rs to builder.rs
Rename flashblocks/payload_builder.rs to flashblocks/builder.rs for
brevity. Update module declaration and import paths in handler, service,
and mod.rs.
